Created getIpcArguments utility.
